requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.10        # Updated to latest working version
celery==5.3.4
redis==5.0.1
pandas==2.1.3
numpy>=1.26.0,<2.0.0           # Required by pandas 2.1.3
tzdata>=2022.1                 # Also required by pandas
openpyxl==3.1.2
gunicorn==21.2.0
python-decouple==3.8
asgiref==3.9.1                 # Required by Django
sqlparse>=0.3.1                # Required by Django
